feat: When Media is "Completed" set all episodes as watched
Updated AddToList button component to handle episodes watched, modified UserSettingsModal CSS, and refactored AnimeEpisodesContainer and MediaPage components
