+++ app/layout/header/components/User/UserSettingsModal/component.module.css
@@ -410,10 +410,10 @@
 /* CHECKBOX */
 #modal .checkbox_container {
 
-    display: flex;
-    flex-direction: row;
+    grid-template-columns: 60px 1fr;
+    display: grid;
+    gap: 16px;
     align-items: center;
-    justify-content: space-between;
 
 }
